Component Extrapolation
- Extrapolated components from single header file.
- Updated tests.
