+++ src/CGravity.cpp
@@ -1,11 +1,11 @@
-#include "../components/Component.h"
+#include "../components/CGravity.h"
 #include "../include/Entity.h"
 
 void CGravity::update(float deltaTime)
 {
     if (owner && owner->getComponent<CTransform>())
     {
-        CTransform *transform = owner->getComponent<CTransform>();
+        auto transform = owner->getComponent<CTransform>();
         transform->velocity += force * deltaTime;
     }
 }
